+++ src/core/eppic.h
@@ -8,6 +8,9 @@
 
 #include "config.h"
 #include "hdf5_wrapper.h"
+#include "numeric.h"
+#include "physical.h"
+#include "vector.h"
 
 /*!
  * World object
@@ -47,11 +50,143 @@ private:
    */
   explicit World(const std::string &input_file_path, const std::string &id);
 
+  /*!
+   * calculates the number of steps required to advance engine state by some
+   * time period
+   * @param adv_t (s) time period to advance by
+   * @return uint64_t
+   */
+  [[nodiscard]] uint64_t calc_num_steps(fpp adv_t) const;
+
+  /*!
+   * calculates the number of steps required to model a given time span
+   * @param time_span (s) time span to be modeled
+   * @return number of steps required by CFL stability condition
+   */
+  [[nodiscard]] uint64_t calc_cfl_steps(fpp time_span) const;
+
+  /*!
+   * advances internal field state by one time step
+   * @param dt (s) time step
+   * @return void
+   */
+  void step(fpp dt);
+
+  /*!
+   * advances internal electric field state by one time step
+   * @param ea electric field a loop constant
+   * @param eb electric field b loop constant
+   */
+  void update_e(fpp ea, fpp eb) const;
+
+  /*!
+   * advances internal magnetic field state by one time step
+   * @param hxa magnetic field a loop constant for x-component
+   * @param hya magnetic field a loop constant for y-component
+   * @param hza magnetic field a loop constant for z-component
+   */
+  void update_h(fpp hxa, fpp hya, fpp hza) const;
+
+  /*!
+   * advances internal electric field x-component state by one time step
+   * @param ea electric field a loop constant
+   * @param eb electric field b loop constant
+   */
+  void update_ex(fpp ea, fpp eb) const;
+  /*!
+   * advances internal electric field y-component state by one time step
+   * @param ea electric field a loop constant
+   * @param eb electric field b loop constant
+   */
+  void update_ey(fpp ea, fpp eb) const;
+
+  /*!
+   * advances internal electric field z-component state by one time step
+   * @param ea electric field a loop constant
+   * @param eb electric field b loop constant
+   */
+  void update_ez(fpp ea, fpp eb) const;
+
+  /*!
+   * advances internal magnetic field x-component state by one time step
+   * @param hya magnetic field a loop constant for y-component
+   * @param hza magnetic field a loop constant for z-component
+   */
+  void update_hx(fpp hya, fpp hza) const;
+
+  /*!
+   * advances internal magnetic field y-component state by one time step
+   * @param hxa magnetic field a loop constant for x-component
+   * @param hza magnetic field a loop constant for z-component
+   */
+  void update_hy(fpp hxa, fpp hza) const;
+
+  /*!
+   * advances internal magnetic field z-component state by one time step
+   * @param hxa magnetic field a loop constant for x-component
+   * @param hya magnetic field a loop constant for y-component
+   */
+  void update_hz(fpp hxa, fpp hya) const;
+
+  /*!
+   * writes internal field state to hdf5 group
+   * @param group HDF5 group to write to
+   */
+  void write_h5(const HDF5Obj &group) const;
+
+  /*!
+   * writes all electric field data to HDF5 group
+   * @param group HDF5 group
+   */
+  void write_h5_e(const HDF5Obj &group) const;
+
+  /*!
+   * writes all magnetic field data to HDF5 group
+   * @param group HDF5 group
+   */
+  void write_h5_h(const HDF5Obj &group) const;
+
+  /*!
+   * writes all electric field x-component data to HDF5 group
+   * @param group HDF5 group
+   */
+  void write_h5_ex(const HDF5Obj &group) const;
+
+  /*!
+   * writes all electric field y-component data to HDF5 group
+   * @param group HDF5 group
+   */
+  void write_h5_ey(const HDF5Obj &group) const;
+
+  /*!
+   * writes all electric field z-component data to HDF5 group
+   * @param group HDF5 group
+   */
+  void write_h5_ez(const HDF5Obj &group) const;
+
+  /*!
+   * writes all magnetic field x-component data to HDF5 group
+   * @param group HDF5 group
+   */
+  void write_h5_hx(const HDF5Obj &group) const;
+
+  /*!
+   * writes all magnetic field y-component data to HDF5 group
+   * @param group HDF5 group
+   */
+  void write_h5_hy(const HDF5Obj &group) const;
+
+  /*!
+   * writes all magnetic field z-component data to HDF5 group
+   * @param group HDF5 group
+   */
+  void write_h5_hz(const HDF5Obj &group) const;
+
   /// configuration from file
   const Config cfg;
 
   /// output HDF5 file
-  const HDF5Obj h5;
+  HDF5Obj h5;
 
   /// xdmf writer
   SimpleXdmf xdmf;
